Added global window stuffs so we don't need to ts-ignore some part of the script and cleaned up ts-ignored (#110)
* added global window stuffs so we don't need to ts-ignore some part of the script

* CI / CD - Prettier Automatic Commit

+++ src/data/sql.cassandra.tsx
@@ -4,13 +4,12 @@ import { getDivider } from 'src/data/sql';
 export function getSelectAllColumns(input: SqlAction.TableInput): SqlAction.Output | undefined {
   const label = `Select All Columns`;
 
-  switch (input.dialect) {
-    case 'cassandra':
-      return {
-        label,
-        formatter: 'sql',
-        query: `SELECT * \nFROM ${input.tableId} \nLIMIT ${input.querySize}`,
-      };
+  if (input.dialect === 'cassandra') {
+    return {
+      label,
+      formatter: 'sql',
+      query: `SELECT * \nFROM ${input.tableId} \nLIMIT ${input.querySize}`,
+    };
   }
 }
 
@@ -26,13 +25,12 @@ export function getSelectSpecificColumns(
   const columnString = `\n` + input.columns.map((col) => `  ${col.name}`).join(',\n');
   const whereColumnString = input.columns.map((col) => `${col.name} = ''`).join('\n -- AND ');
 
-  switch (input.dialect) {
-    case 'cassandra':
-      return {
-        label,
-        formatter: 'sql',
-        query: `SELECT ${columnString} \nFROM ${input.tableId} \n -- WHERE ${whereColumnString} \nLIMIT ${input.querySize}`,
-      };
+  if (input.dialect === 'cassandra') {
+    return {
+      label,
+      formatter: 'sql',
+      query: `SELECT ${columnString} \nFROM ${input.tableId} \n -- WHERE ${whereColumnString} \nLIMIT ${input.querySize}`,
+    };
   }
 }
 
@@ -46,13 +44,12 @@ export function getInsertCommand(input: SqlAction.TableInput): SqlAction.Output 
   const columnString = input.columns.map((col) => col.name).join(',\n');
   const insertValueString = input.columns.map((col) => `'_${col.name}_'`).join(',\n');
 
-  switch (input.dialect) {
-    case 'cassandra':
-      return {
-        label,
-        formatter: 'sql',
-        query: `INSERT INTO ${input.tableId} (\n${columnString}\n) VALUES (\n${insertValueString}\n)`,
-      };
+  if (input.dialect === 'cassandra') {
+    return {
+      label,
+      formatter: 'sql',
+      query: `INSERT INTO ${input.tableId} (\n${columnString}\n) VALUES (\n${insertValueString}\n)`,
+    };
   }
 }
 
@@ -66,13 +63,12 @@ export function getUpdateCommand(input: SqlAction.TableInput): SqlAction.Output 
   const columnString = input.columns.map((col) => `-- ${col.name} = ''`).join(',\n');
   const whereColumnString = input.columns.map((col) => `-- ${col.name} = ''`).join(' AND \n');
 
-  switch (input.dialect) {
-    case 'cassandra':
-      return {
-        label,
-        formatter: 'js',
-        query: `UPDATE ${input.tableId}\n SET \n${columnString}\n WHERE ${whereColumnString}`,
-      };
+  if (input.dialect === 'cassandra') {
+    return {
+      label,
+      formatter: 'js',
+      query: `UPDATE ${input.tableId}\n SET \n${columnString}\n WHERE ${whereColumnString}`,
+    };
   }
 }
 
@@ -85,13 +81,12 @@ export function getDeleteCommand(input: SqlAction.TableInput): SqlAction.Output 
 
   const whereColumnString = input.columns.map((col) => `-- ${col.name} = ''`).join(' AND \n');
 
-  switch (input.dialect) {
-    case 'cassandra':
-      return {
-        label,
-        formatter: 'sql',
-        query: `-- DELETE FROM ${input.tableId} \n -- WHERE\n ${whereColumnString}`,
-      };
+  if (input.dialect === 'cassandra') {
+    return {
+      label,
+      formatter: 'sql',
+      query: `-- DELETE FROM ${input.tableId} \n -- WHERE\n ${whereColumnString}`,
+    };
   }
 }
